pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "langgraph-supervisor"
version = "0.0.2"
description = "An implementation of a supervisor multi-agent architecture using LangGraph"
authors = [
    {name = "Vadym Barda", email = "19161700+vbarda@users.noreply.github.com "}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.17",
    "langchain-openai>=0.3.5",
    "langgraph>=0.2.71",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "tavily-python>=0.5.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]
